# data-api/open-api-specification.yml
openapi: '3.0.0'
info:
  title: 'Equinity-data-api'
  version: '1.0.0'
paths:
  /api/internal/v1/equinity/search/:
    get:
      tags:
        - Search
      summary: 'Search CRM forms'
      operationId: 'doSearchBy'
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/searchUsn'
        - $ref: '#/components/parameters/searchType'
        - $ref: '#/components/parameters/searchClient'
        - $ref: '#/components/parameters/searchClientDoB'
        - $ref: '#/components/parameters/searchSubmittedFrom'
        - $ref: '#/components/parameters/searchSubmittedTo'
        - $ref: '#/components/parameters/searchProviderAccount'
        - $ref: '#/components/parameters/searchPage'
        - $ref: '#/components/parameters/searchPageSize'
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResultDTO"
        '400':
          description: 'Bad request'
        '403':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

  /api/internal/v1/equinity/crm5/{usn}:
    get:
      tags:
        - CRM5Interface
      operationId: getApplication
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/crmUsn'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crm5FormDTO'

  /api/internal/v1/equinity/crm4/{usn}:
    get:
      tags:
        - CRM4Interface
      operationId: getApplicationCrm4
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/crmUsn'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crm4FormDTO'

  /api/internal/v1/equinity/crm7/{usn}:
    get:
      tags:
        - CRM7Interface
      operationId: getApplicationCrm7
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/crmUsn'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crm7FormDTO'

  /api/internal/v1/equinity/crm14/{usn}:
    get:
      tags:
        - CRM14Interface
      operationId: getApplicationCrm14
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/crmUsn'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crm14FormDTO'
  ###
  # Reports
  ###
  /api/internal/v1/equinity/report/crm4/{decisionFrom}/{decisionTo}:
    get:
      tags:
        - ReportCrm4
      summary: 'CRM4 Periodical Report'
      operationId: 'generateReportCrm4'
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/reportDecisionFrom'
        - $ref: '#/components/parameters/reportDecisionTo'
      responses:
        '200':
          description: 'Success'
          content:
            application/octet-stream:
              schema:
                type: string
        '400':
          description: 'Bad request'
        '403':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /api/internal/v1/equinity/report/crm5/{decisionFrom}/{decisionTo}:
    get:
      tags:
        - ReportCrm5
      summary: 'CRM5 Periodical Report'
      operationId: 'generateReportCrm5'
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - $ref: '#/components/parameters/reportDecisionFrom'
        - $ref: '#/components/parameters/reportDecisionTo'
      responses:
        '200':
          description: 'Success'
          content:
            application/octet-stream:
              schema:
                type: string
        '400':
          description: 'Bad request'
        '403':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /api/internal/v1/equinity/report/crm14/:
    get:
      tags:
        - ReportCrm14
      summary: 'CRM14 Summary Report'
      operationId: 'generateReportCrm14'
      parameters:
        - $ref: '#/components/parameters/profileAcceptedTypes'
        - name: filterByDecision
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByFlag'
        - name: decisionFrom
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: decisionTo
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: filterBySubmit
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByFlag'
        - name: submittedFrom
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: submittedTo
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: filterByCreation
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByFlag'
        - name: createdFrom
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: createdTo
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: filterByLastSubmit
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByFlag'
        - name: lastSubmittedFrom
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - name: lastSubmittedTo
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ParamFilterByDate'
        - $ref: '#/components/parameters/reportState'

      responses:
        '200':
          description: 'Success'
        '400':
          description: 'Bad request'
        '403':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

components:
  parameters:
    crmUsn:
      name: usn
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 2008372910
    searchUsn:
      name: usn
      in: query
      required: false
      schema:
        type: string
        example: '3456'
        minLength: 3
        maxLength: 10
        pattern: '[0-9]{3,10}'
    searchType:
      name: type
      description: 'CRM form Type'
      in: query
      required: false
      schema:
        type: integer
        example: 4
        minimum: 0
        maximum: 999
    searchClient:
      name: client
      description: 'Client name'
      in: query
      required: false
      schema:
        type: string
        example: 'John Doe'
        minLength: 3
        maxLength: 256
    searchClientDoB:
      name: clientDoB
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ParamFilterByDate'
    searchSubmittedFrom:
      name: submittedFrom
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ParamFilterByDate'
    searchSubmittedTo:
      name: submittedTo
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ParamFilterByDate'
    searchProviderAccount:
      name: providerAccount
      in: query
      required: false
      schema:
        type: string
        minLength: 4
        maxLength: 6
        pattern: "([0-9A-Za-z]){4,6}"
        example: '0X12YZ'
    searchPage:
      name: page
      in: query
      required: false
      schema:
        default: 0
        type: integer
        minimum: 0
        example: 0
    searchPageSize:
      name: pageSize
      in: query
      required: false
      schema:
        default: 10
        type: integer
        minimum: 10
        maximum: 100
        example: 20
    profileAcceptedTypes:
      name: profileAcceptedTypes
      in: header
      required: false
      schema:
        type: string
        example: 1,2,3

    reportDecisionFrom:
      name: decisionFrom
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ParamFilterByDate'
    reportDecisionTo:
      name: decisionTo
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ParamFilterByDate'
    reportState:
      name: state
      in: query
      required: false
      schema:
        type: string
        minLength: 2
        maxLength: 40
        default: All
        example: All

  schemas:
    ParamFilterByDate:
      type: string
      pattern: "^[0-9]{4}-(1[0-2]|0[1-9])-(3[01]|[12][0-9]|0[1-9])$"
      example: '2019-04-13'
    ParamFilterByFlag:
      type: integer
      minimum: 0
      maximum: 1
      example: '0|1'

    ## Search
    SearchResultDTO:
      $ref: "./schemas/search.yml#/SearchResultDTO"

    ## CRM 5
    Crm5FormDTO:
      type: object
      properties:
        formDetails:
          $ref: "./schemas/crm5.yml#/CRM5DetailsDTO"
        evidenceFiles:
          $ref: '#/components/schemas/CrmEvidenceFilesDTO'

    ## CRM 4
    Crm4FormDTO:
      type: object
      properties:
        formDetails:
          $ref: "./schemas/crm4.yml#/Crm4DetailsDTO"
        evidenceFiles:
          $ref: '#/components/schemas/CrmEvidenceFilesDTO'

    ## CRM 7
    Crm7FormDTO:
      type: object
      properties:
        formDetails:
          $ref: "./schemas/crm7.yml#/Crm7DetailsDTO"
        evidenceFiles:
          $ref: '#/components/schemas/CrmEvidenceFilesDTO'

    ## CRM 14
    Crm14FormDTO:
      type: object
      properties:
        formDetails:
          $ref: "./schemas/crm14.yml#/Crm14DetailsDTO"
        evidenceFiles:
          $ref: '#/components/schemas/CrmEvidenceFilesDTO'

#    ## Linked Attachments
    CrmEvidenceFilesDTO:
      $ref: "./schemas/crmEvidenceFiles.yml#/CrmEvidenceFilesDTO"